ml/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.9.0.80
mediapipe==0.10.14
joblib==1.3.2
numpy==1.26.3
pillow==10.2.0
scikit-learn==1.4.0
ultralytics>=8.0.0
